rewrite data models and introduce json data parser

+++ src/sk/upjs/ics/diplomovka/data/stands/closures/WingspanClosureCondition.java
@@ -4,9 +4,9 @@ import sk.upjs.ics.diplomovka.data.flights.Flight;
 
 public class WingspanClosureCondition implements ClosureCondition {
 
-    private int wingspan;
+    private double wingspan;
 
-    public WingspanClosureCondition(int wingspan) {
+    public WingspanClosureCondition(double wingspan) {
         this.wingspan = wingspan;
     }
